headerCheck: Repair headers if ANY are valid
Before this commit, repairing was only possible if the first header was valid. Now any one header can be valid, and it will be restored to the others.

This won't really change much, but it's more effective since all 4 checksums have to be invalid for a file to be invalid.