Fast and succinct parser
With this update, my parsing function is damn near perfect.

One of the purposes of parsing is to check for any errors in the filesystem. A valid checksum alone does not guarantee that the data is intact. To accomplish this, it reads the data in the same way that the N64 would, and finds situations that would hinder, prevent or otherwise obscure the typical reading process.

My earlier parsing method was through a combination of functions, in a rather haphazard implementation. It works, but it wasn't clear just how it worked. This new version streamlines the logic into a single function, strips unnecessary logic and function calls, making it faster and more accurate. I also changed some of the error detection conditions to ones that seem to cover more issues. So I am pretty confident that it should catch any possible error. However it won't allow two notes using the same index, whereas on a real N64 it seems to allow this. I highly doubt that it would ever occur in the wild though and not a hacked MPK file.

This iteration of the parser does not "repair" files at all, and does not differentiate between primary/backup (yet).
I either have to: a) Detect and restore backup data to primary, or: b) Wipe both primary/backup and overwrite with the valid data. Option b would follow more closely to the N64, and could effectively repair "unassigned" indexes, however they shouldn't be there in the first place. Option a would be faster, retaining the original data (which my parser already knows is valid).

TODO: 
1. Target selected IndexTable (variable o) - and copy it to the other one. (Check / Repair Index checksum)
2. Wipe empty/invalid notes clean (Repair 0x08:2!!!!).
3. Implement header checksum checking (and repair!)
4. Implement edit interface (View, Save, Import/Export, Delete)
2. Decode note metadata, and output struct nicely:
number id[]:
    array indexes
    string noteName
    string gameCode
    string publisherCode
etc...