Many changes. Parser is done.
Took the parser back to the drawing board, looking at each step of the parsing algorithm. Turned out to be essentially identical to what was already implemented, so I ended up only refining the code. A big addition was header checking as well as data repair. Now, when parsing is successful, it will repair header checksums, as well as the indexTable checksum. Then it restores or backs up the valid indexTable, and fixes missing bits in the noteTable.  Lastly, I added noteName decoding. 

So in short, the parsing is pretty complete (not to mention isolated in a single function) beside some obligatory code cleanups and other testing. The change in design means I will probably have to rewrite the functions for importing/exporting and deleting notes.