Fixes and whatnot
Entire page is no longer rebuilt on UI updates. Only the table rows are
rebuilt, and other targets are individualized. topToolbar is now static
in the HTML.

Separated CSS from the HTML, and made the fonts offline.

A bunch of fixes and modifications to the main script.
Event bugs fixed in Firefox.
Output filename extension is replaced with MPK.