Chrome app: Save As updates filename
Within the Chrome app, when you press 'Save As', you are able to choose
a name to save the file as.

The original behaviour would be to keep the original file handle - using
'Save As' only allowed you to create a copy. However, now it updates the
file handle, meaning when you press Save normally, it updates the newly
created file. The only problem, is that the UI didn't update accordingly
when this occured, so now it does.