Use this keyword + Implement NoteTable reorder
*NoteTable reorder*: in readNotes, I added a routine that collects
present data and overwrites the noteTable. It effectively removes gaps
between Notes in the table, which eliminates issues when deleting notes,
or (soon to be added), reordering notes.

I removed hard-coded references to namespaces, replacing them with the
`this` keyword. I am already using `bind()` on every event handler, and
they are the only that would cause an issue.

Next: Internalizing the app init code.