Implement comment support
With this commit, comment support should be fully working for DexDrive
import, note file import, and note file export. As well as general
editing of comments. For now, comments cannot be saved to an MPK file
(i'll figure something out later).

No unicode support. Only 256 byte characters in ISO-8859-1 encoding
(basically first page of UTF-8). This should support most latin-based
languages... but if I can figure out how to do UTF-8 properly, I'll use
that instead.

In order to make this work, I had to add an extension (a header) to note
files. It is a bit of a prototype, but could very well be the basis for
a proper finalized note file format. So note files made today should
hopefully work in the future.