Add note reorder mode (finally)
I originally wanted to do this in 2016, but wasn't happy with the way it worked back then. This time it works quite well and is essentially lossless and noninvasive because it just 'swaps' two rows rather than 'shift' and resort the entire list.

It is enabled through the settings menu.