Add sumCheck for NoteTable
In 598 MPK files, all note data within contain 0000 in these bytes so we can use it for parsing. In total, that is 9,568 checks passed, so I am confident this behavior is reliable. This value is written to 0 each time a game saves.

There is only 1 MPK file that doesn't pass the check, but it's just leftover garbage data, so it can be ignored.