Better note code checking + Entry Check
Previously, if a note was found but its game code values were invalid, the file would be rejected due to a mismatch. However there may still be usable data on the file.

This fix will try to catch this scenario and apply a fix and display an error, but otherwise attempt to load the file.