Simulate "official" ID serial generation
I don't have enough data to conclusively determine the original format. This is just a best guess. In fact it may be better to just keep the original randomized 24 byte scheme that InterAct used originally. But it'd be nice to get as close as possible to original hardware.

Also, committing some old debug code that's been sitting around for a while.