Fast Index check before terminating parser
In rare cases all ID checksums can fail, yet the Index table is intact.
N64 hardware recovers from this state by asking the user to "Repair ID".

Previously, MPKEdit would reject any file with an invalid header.
But since this is not what N64 hardware does, we should do better.

I also don't really want to give vague generic prompts to the user.
What we do is we do a fast parse of the Index table if the header is invalid, and based on that, we decide whether to truly terminate the parser.

This solution should be relatively efficient.
But it does not yet actually repair the ID. I'll do that later.