Preliminary IndexTable repair support
If an error is found in the primary IndexTable, the backup is checked. If the backup has no errors, it will be used. If the backup does have errors, it normally means the data cannot be loaded.

But it is possible to isolate the corruption down to only the affected notes. This could allow the file to still be able to load the valid data, while simply erasing the corrupt data.

The code for this is experimental, and does not cover all potential situations, but it seems promising. More testing will be needed.